Broke sbrk(). First attempt implementing lazy page allocation, causes 'panic: apicid'.

+++ trap.c
@@ -32,6 +32,54 @@ idtinit(void)
   lidt(idt, sizeof(idt));
 }
 
+
+// JTM - Stole walkpgdir() from vm.c
+static pte_t *
+walkpgdir(pde_t *pgdir, const void *va, int alloc)
+{
+  pde_t *pde;
+  pte_t *pgtab;
+
+  pde = &pgdir[PDX(va)];
+  if(*pde & PTE_P){
+    pgtab = (pte_t*)P2V(PTE_ADDR(*pde));
+  } else {
+    if(!alloc || (pgtab = (pte_t*)kalloc()) == 0)
+      return 0;
+    // Make sure all those PTE_P bits are zero.
+    memset(pgtab, 0, PGSIZE);
+    // The permissions here are overly generous, but they can
+    // be further restricted by the permissions in the page table
+    // entries, if necessary.
+    *pde = V2P(pgtab) | PTE_P | PTE_W | PTE_U;
+  }
+  return &pgtab[PTX(va)];
+}
+
+// JTM - Stole mappages() from vm.c
+static int
+mappages(pde_t *pgdir, void *va, uint size, uint pa, int perm)
+{
+  char *a, *last;
+  pte_t *pte;
+
+  a = (char*)PGROUNDDOWN((uint)va);
+  last = (char*)PGROUNDDOWN(((uint)va) + size - 1);
+  for(;;){
+    if((pte = walkpgdir(pgdir, a, 1)) == 0)
+      return -1;
+    if(*pte & PTE_P)
+      panic("remap");
+    *pte = pa | perm | PTE_P;
+    if(a == last)
+      break;
+    a += PGSIZE;
+    pa += PGSIZE;
+  }
+  return 0;
+}
+
+
 //PAGEBREAK: 41
 void
 trap(struct trapframe *tf)
@@ -86,12 +134,29 @@ trap(struct trapframe *tf)
               tf->trapno, cpuid(), tf->eip, rcr2());
       panic("trap");
     }
+   
+    //JTM - Check if the error was a page fault
+    
+    if(tf->trapno == 14){
+	// Allocate page lazily
+	char* mem;
+
+	if((mem = kalloc()) == 0){
+		panic("Trap 14: kalloc() returned 0 before mappages().");
+	}
+
+	mappages(myproc()->pgdir, (void*)rcr2(), myproc()->sz, V2P(mem), PTE_W|PTE_U);
+
+	break;
+    }
+    
     // In user space, assume process misbehaved.
     cprintf("pid %d %s: trap %d err %d on cpu %d "
             "eip 0x%x addr 0x%x--kill proc\n",
             myproc()->pid, myproc()->name, tf->trapno,
             tf->err, cpuid(), tf->eip, rcr2());
     myproc()->killed = 1;
+    
   }
 
   // Force process exit if it has been killed and is in user space.
